Added PagerTabStrip to schedule screen.
The PagerTabStrip library has been added.

in SchedulePagerTabStripViewController.swift
that controller is for adding view controllers to the PagerTabStrip of the Schedule ViewController.
set the style settings of the ButtonBarItem and adding viewcontrollers to the PagerTabStrip.

RoomEventsViewController.swift is the ViewController,
that will contain the content of each screen
in the Schedule PagerTabStrip ViewController.