Main Topics: Added add event button in SchedulePagerTabStripViewController. Bug fix in Edit UIBarButtonItem in ProfileViewController, for highlighted state. Added 'Add Events' image set to assets.
in SchedulePagerTabStripViewController.swift:
added subclass of UIButton called ButtonWithImage

in Main.storyboard:
added UIButton for add events that type of ButtonWithImage class.
added segue from ProfileTableViewController to SetEditProfileViewController with 'EditProfile' id

in ProfileViewController:
deleted the ProfileViewController class
renamed the file to ProfileTableViewController.swift
added the bug fix in UIBarButtonItem highlighted state to be with OpenSansHebrew-Bold font.
on prepare for segue with 'EditProfile' id to SetEditProfileViewController, i set the profileMode to edit mode.