adding event image from server to EventTableViewController
in Main.storyboard:
changing image view cell to be 44 height by default.

in EventTableViewController.swift:
deleted image property
adding to initiate function, call to image url and putting the image returned to eventImageView.
adding dynamic height for the cell with the image view to be relative to the ratio of the image.